mostly smaller changes
Some texts changed.
The order of G is displayed in the "Select Secret" dialog if the curve
size is large.
The animation, transitioning the binary code from Alice to Bob, adapts
to the size of the canvas.
key.png now has an alpha channel -> is transparent.
Display location of the key changed slightly.
Addresses #117 and #33. In #33 all issues are solved.
Removed redundant redraw() commands.
Problem: The animation stops on the way from Bob to Alice, was
previously so.