Several improvements
Fixes Problem that some step results are not displayed or skipped.
Fixes Problem that causes the first step to be skipped.
Fixes Problem with the table displaying the current values not
displaying them correctly.
Removed surpressWarning Tags from Code.
Added show info dialogs checkbox to the options group.
Repaired the possibility to press Encrypt or Sign directly at the
beginning that the calculations are done with default values.
Added toString() method to RSAData.
Some minor text changes.
In RSAView the setFocus() is set to the parent composite.