various GUI improvements
Added the c to the setFocus() to enable direct scrolling.
Removed the unnecessary TabItem.
Renamed many variables to fit the java naming conventions.
Fixed the label that display whether a verification was successfull or
not.
Removed paintListerners from load Buttons.
Added code to the selectionListener of the pki button to open the pki
plugin.