GUI improvements, general bug fixing and cleaning of the HE plug-in code
My main goal was to fix the GUI for usage with up-scaled high-DPI
displays. Therefore, I replaced everything that was not in GridLayout
with GridLayout.

I noticed some bugs in the plug in code, which I fixed, too.

One of them was a infinite while-loop, running in a new thread when a
keyboard entry was made in a certain text field.

Another one was conerning the loading of password-secured key files. In
the Gentry/Halevi and in the Paillier views, entering a wrong password
and then pressing "Finish" resulted in a) an exception logged to the
console and b) closing the dialog window and loading the keys as if the
password would have been correct.

Also, there were some errors regarding the correct output of error
messages in the titleArea of the wizards.

Last but not least, there was a lot of duplicate code (duplicate files,
really), which I deleted. All of the WizardPage subclasses existed 2
times, once in the he.wizards package and once in the he.wizards.pages
package.