Removed the hex-value textfield and replaced the plaintext textfield with a NumberblocksAndTextViewer.
Some smaller changes:
- Errors will be logged in ErrorLog
- plaintext, ciphertext and signature are stored in List<integer>
Revised the Key Entry Wizard

- added proper exception handling
- commented the code for future adaptions
- small usability improvements like automatic insert old data.

Externalized Strings
Added Info Dialogs

When inheriting data from another tab it was possible to load keys from
the key store that does not exist. This is not possible anymore. The
user gets feedback that it does not work in form of a dialog.

When inheriting a key that exists and the user has entered a wrong
password the inherit function fails. This is fixed now. A Dialog opens
saying the password was probably wrong and the enter password dialog
opens again.
Added comments that make the code easier to understand