Issue#376 [UI] Allow focus-switching in node pop-up
Previously the FlowChartNode popup closed on focus loss on ANY of its children.
This was a problem when a user would just switch focus to another child
of the popup.

Add logic to detect those focus switches (sadly SWT does not provide anything
out-of-the-box) and allow the user to freely navigate in the pop-up

Additionally, improve the "accidental re-open" code, when a user directly
clicks on the button, so both focusLost --> close and buttonPressed --> open
are triggered.