[Algorithm] Fix all key re-creation "bugs"
In fact they were not bugs but rather missing logic in the calling code.
Previously the sessions were initialized as if both entities were
running on the same machine (which in fact they are).

But this was not clean and lead to some problems.
Disentangle all that logic (and more or less rewrite it completely).

Summary of changes:

* Add a "server" through which keys are exchanged.
* Remove confusing classes such as SessionInitialization and
  PreSessionparameter
* Properly store all keys of the entities in the InMemoryStores
* Properly initialize the decrypting session cipher from the pre-key
  message (This was not done before, missing the point of the pre-key
  message).
* Improve code quality in general